+++ py-polars/tests/unit/test_df.py
@@ -14,8 +14,8 @@ import pyarrow as pa
 import pytest
 
 import polars as pl
+from polars._construction import iterable_to_pydf
 from polars.datatypes import DTYPE_TEMPORAL_UNITS, INTEGER_DTYPES
-from polars.internals.construction import iterable_to_pydf
 from polars.testing import (
     assert_frame_equal,
     assert_frame_not_equal,
